m2sdr_sata: move catalog helpers out of CLI
Split the named capture catalog data model, parser, formatter, validation and allocation helpers into m2sdr_sata_catalog.c/.h.

Keep the SATA host-buffer read/write wrappers in m2sdr_sata.c so this change preserves the existing I/O paths while reducing the size of the command implementation.

Validated with make m2sdr_sata and ./m2sdr_sata --help.
